[FLINK-23061] Split BootstrapTools
