+++ flink-runtime/src/main/java/org/apache/flink/runtime/rpc/akka/AkkaRpcServiceUtils.java
@@ -22,7 +22,6 @@ import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.configuration.AkkaOptions;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.runtime.akka.AkkaUtils;
-import org.apache.flink.runtime.clusterframework.BootstrapTools;
 import org.apache.flink.runtime.highavailability.HighAvailabilityServicesUtils;
 import org.apache.flink.runtime.highavailability.HighAvailabilityServicesUtils.AddressResolution;
 import org.apache.flink.runtime.net.SSLUtils;
@@ -281,8 +280,8 @@ public class AkkaRpcServiceUtils {
         private String actorSystemName = AkkaUtils.getFlinkActorSystemName();
 
         @Nullable
-        private BootstrapTools.ActorSystemExecutorConfiguration actorSystemExecutorConfiguration =
-                null;
+        private AkkaBootstrapTools.ActorSystemExecutorConfiguration
+                actorSystemExecutorConfiguration = null;
 
         @Nullable private Config customConfig = null;
         private String bindAddress = NetUtils.getWildcardIPAddress();
@@ -317,7 +316,7 @@ public class AkkaRpcServiceUtils {
         }
 
         public AkkaRpcServiceBuilder withActorSystemExecutorConfiguration(
-                final BootstrapTools.ActorSystemExecutorConfiguration
+                final AkkaBootstrapTools.ActorSystemExecutorConfiguration
                         actorSystemExecutorConfiguration) {
             this.actorSystemExecutorConfiguration = actorSystemExecutorConfiguration;
             return this;
@@ -349,7 +348,7 @@ public class AkkaRpcServiceUtils {
                 throws Exception {
             if (actorSystemExecutorConfiguration == null) {
                 actorSystemExecutorConfiguration =
-                        BootstrapTools.ForkJoinExecutorConfiguration.fromConfiguration(
+                        AkkaBootstrapTools.ForkJoinExecutorConfiguration.fromConfiguration(
                                 configuration);
             }
 
@@ -358,7 +357,7 @@ public class AkkaRpcServiceUtils {
             if (externalAddress == null) {
                 // create local actor system
                 actorSystem =
-                        BootstrapTools.startLocalActorSystem(
+                        AkkaBootstrapTools.startLocalActorSystem(
                                 configuration,
                                 actorSystemName,
                                 logger,
@@ -367,7 +366,7 @@ public class AkkaRpcServiceUtils {
             } else {
                 // create remote actor system
                 actorSystem =
-                        BootstrapTools.startRemoteActorSystem(
+                        AkkaBootstrapTools.startRemoteActorSystem(
                                 configuration,
                                 actorSystemName,
                                 externalAddress,
